chore: 패키지 이동 및 depth 조정
