+++ README.md
@@ -2,15 +2,14 @@
 
 ## 주요 요구 사항 🚀
 
-- [ ] 단위 테스트하기 어려운 코드와 단위 테스트 가능한 코드를 분리해 단위 테스트 가능한 코드에 대해 단위 테스트를 구현한다.
-- [ ] Spring Data JPA 사용 시 spring.jpa.hibernate.ddl-auto=validate 옵션을 필수로 준다.
-- [ ] 데이터베이스 스키마 변경 및 마이그레이션이 필요하다면 아래 문서를 적극 활용한다.
-    - [ ] DB도 형상관리를 해보자!
-- [ ] 자바 코드 컨벤션을 지키면서 프로그래밍한다.
-    - [ ] 단, 들여쓰기는 '2 spaces'가 아닌 '4 spaces'로 한다.
+- [x] 단위 테스트하기 어려운 코드와 단위 테스트 가능한 코드를 분리해 단위 테스트 가능한 코드에 대해 단위 테스트를 구현한다.
+- [x] 데이터베이스 스키마 변경 및 마이그레이션이 필요하다면 아래 문서를 적극 활용한다.
+    - [x] DB도 형상관리를 해보자!
+- [x] 자바 코드 컨벤션을 지키면서 프로그래밍한다.
+    - [x] 단, 들여쓰기는 '2 spaces'가 아닌 '4 spaces'로 한다.
 - [ ] indent(인덴트, 들여쓰기) depth를 2를 넘지 않도록 구현한다. 1까지만 허용한다.
-- [ ] 모든 기능을 TDD로 구현해 단위 테스트가 존재해야 한다. 단, UI(System.out, System.in) 로직은 제외
-- [ ] 기타 원칙 지키기
+- [x] 모든 기능을 TDD로 구현해 단위 테스트가 존재해야 한다. 단, UI(System.out, System.in) 로직은 제외
+- [x] 기타 원칙 지키기
     - 핵심 로직을 구현하는 코드와 UI를 담당하는 로직을 구분한다.
     - UI 로직을 InputView, ResultView와 같은 클래스를 추가해 분리한다.
     - 함수(또는 메서드)의 길이가 10라인을 넘어가지 않도록 구현한다.
