+++ src/main/java/kitchenpos/dao/TableGroupDao.java
@@ -1,13 +1,12 @@
 package kitchenpos.dao;
 
 import java.util.List;
-import java.util.Optional;
 import kitchenpos.domain.TableGroup;
 
 public interface TableGroupDao {
     TableGroup save(TableGroup entity);
 
-    Optional<TableGroup> findById(Long id);
+    TableGroup findById(Long id);
 
     List<TableGroup> findAll();
 }
