+++ src/main/java/kitchenpos/dao/jdbctemplate/JdbcTemplateTableGroupDao.java
@@ -1,13 +1,15 @@
-package kitchenpos.dao;
+package kitchenpos.dao.jdbctemplate;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.TableGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
@@ -15,14 +17,14 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class JdbcTemplateMenuDao{
-    private static final String TABLE_NAME = "menu";
+public class JdbcTemplateTableGroupDao {
+    private static final String TABLE_NAME = "table_group";
     private static final String KEY_COLUMN_NAME = "id";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public JdbcTemplateMenuDao(final DataSource dataSource) {
+    public JdbcTemplateTableGroupDao(final DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -30,16 +32,13 @@ public class JdbcTemplateMenuDao{
         ;
     }
 
-    public Menu save(final Menu entity) {
-        final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("name", entity.getName())
-                .addValue("price", entity.getPrice())
-                .addValue("menu_group_id", entity.getMenuGroupId());
+    public TableGroup save(final TableGroup entity) {
+        final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         final Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
-    public Optional<Menu> findById(final Long id) {
+    public Optional<TableGroup> findById(final Long id) {
         try {
             return Optional.of(select(id));
         } catch (final EmptyResultDataAccessException e) {
@@ -47,30 +46,21 @@ public class JdbcTemplateMenuDao{
         }
     }
 
-    public List<Menu> findAll() {
-        final String sql = "SELECT id, name, price, menu_group_id FROM menu ";
+    public List<TableGroup> findAll() {
+        final String sql = "SELECT id, created_date FROM table_group";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    public long countByIdIn(final List<Long> ids) {
-        final String sql = "SELECT COUNT(*) FROM menu WHERE id IN (:ids)";
-        final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("ids", ids);
-        return jdbcTemplate.queryForObject(sql, parameters, Long.class);
-    }
-
-    private Menu select(final Long id) {
-        final String sql = "SELECT id, name, price, menu_group_id FROM menu WHERE id = (:id)";
+    private TableGroup select(final Long id) {
+        final String sql = "SELECT id, created_date FROM table_group WHERE id = (:id)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
                 .addValue("id", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private Menu toEntity(final ResultSet resultSet) throws SQLException {
-        return new Menu(resultSet.getLong("id"),
-                resultSet.getString("name"),
-                resultSet.getBigDecimal("price"),
-                resultSet.getLong("menu_group_id"),
+    private TableGroup toEntity(final ResultSet resultSet) throws SQLException {
+        return new TableGroup(resultSet.getLong(KEY_COLUMN_NAME),
+                resultSet.getObject("created_date", LocalDateTime.class),
                 new ArrayList<>());
     }
 }
