+++ src/main/java/kitchenpos/dao/repository/TableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.dao.repository;
 
 import java.util.List;
-import kitchenpos.dao.JdbcTemplateOrderTableDao;
+import kitchenpos.dao.jdbctemplate.JdbcTemplateOrderTableDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.domain.OrderTable;
 import org.springframework.dao.InvalidDataAccessApiUsageException;
